Fix three type conversion warnings
- Fix CGFloat to float conversion in PBGitHistoryController.m:415 (dividerThickness)
- Fix CGFloat to float conversion in PBGitHistoryController.m:438 (frame height)
- Fix int to NSUInteger conversion in OpenRecentController.m:107 (index for NSIndexSet)

+++ Classes/Controllers/OpenRecentController.m
@@ -104,7 +104,7 @@
 			int index = (int)[currentResults indexOfObject: selectedResult]-1;
 			if(index < 0) index = 0;
 			selectedResult = [currentResults objectAtIndex:(NSUInteger)index];
-			[resultViewer selectRowIndexes:[NSIndexSet indexSetWithIndex:index] byExtendingSelection:FALSE];
+			[resultViewer selectRowIndexes:[NSIndexSet indexSetWithIndex:(NSUInteger)index] byExtendingSelection:FALSE];
 			[resultViewer scrollRowToVisible:index];
 		}
 		result = YES;
